Change the data type of NLS to float.
In the UKM data, I found there are 1e-5% rows where NLS > CLS and each of them is NLS = CLS + 1.
I suspected this was caused by the precision differences between double and float.
CLS is a float number while the type of NLS is double.
This CL is to convert NLS to a float number.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2905995
Commit-Queue: Hongbo Song <hbsong@google.com>
Cr-Commit-Position: refs/heads/master@{#884775}

+++ chrome/browser/page_load_metrics/observers/back_forward_cache_page_load_metrics_observer_browsertest.cc
@@ -537,7 +537,7 @@ IN_PROC_BROWSER_TEST_F(BackForwardCachePageLoadMetricsObserverBrowserTest,
             content::RenderFrameHost::LifecycleState::kInBackForwardCache);
 
   // Go back to A.
-  double next_score;
+  float next_score;
   {
     auto waiter = CreatePageLoadMetricsTestWaiter();
     waiter->AddPageExpectation(
